# docs/swagger.yaml
definitions:
  commun.MonitorChart2:
    properties:
      title:
        description: title
        type: string
      x:
        description: X-axis, time data
        items:
          type: string
        type: array
      "y":
        description: Y-axis, data
        items:
          type: string
        type: array
    type: object
  commun.WarnChart:
    properties:
      column:
        $ref: '#/definitions/commun.MonitorChart2'
      pie:
        items:
          $ref: '#/definitions/commun.WarnPieItem'
        type: array
    type: object
  commun.WarnPieItem:
    properties:
      count:
        type: integer
      name:
        type: string
      percentage:
        type: number
    type: object
  handle.Threshold:
    properties:
      cpu:
        type: integer
      disk:
        type: integer
      memory:
        type: integer
    type: object
...
  notice.DingTalk:
    properties:
      agentID:
        type: string
      appKey:
        type: string
      appSecret:
        type: string
      hookAddr:
        type: string
      hookSecret:
        type: string
      hookTitle:
        type: string
      receive:
        type: string
      sessionID:
        type: string
    type: object
  notice.PushConfig:
    properties:
      dingTalk:
        $ref: '#/definitions/notice.DingTalk'
      smtp:
        $ref: '#/definitions/notice.SMTP'
      weChat:
        $ref: '#/definitions/notice.WeChat'
    type: object
  notice.SMTP:
    properties:
      addr:
        type: string
      pass:
        type: string
      port:
        type: integer
      receiveEmail:
        type: string
      ssl:
        type: boolean
      topic:
        type: string
      user:
        type: string
    type: object
  notice.WeChat:
    properties:
      eappID:
        type: integer
      eappVoucher:
        type: string
      eid:
        type: string
      receiveUserID:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/bigscreen/3:
    get:
      parameters:
      - description: ' time type, 0 1hour, 1 1day, 2 7day '
        in: formData
        name: type
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/commun.WarnChart'
      summary: Warn big screen chart
      tags:
      - bigscreen
...
  /api/v1/download/download_log:
    get:
      parameters:
      - description: ' size'
        in: query
        name: size
        type: string
      summary: log download
      tags:
      - download
...
  /api/v1/monitor/nodeMonitor:
    get:
      parameters:
      - description: chart type, 0 Cpu usage, 1 Cpu load, 2 Mem rate, 3 Disk part
          rate, 4 Network recv, 5 Network send
        in: query
        name: chartType
        required: true
        type: string
      - description: time type, 0 real time, 1 24hour, 2 7days, 3 custom
        in: query
        name: timeType
        required: true
        type: string
      - description: time granular, 0 1minute, 1 5minute, 2 15minute, 3 1hour, 4 1day
        in: query
        name: granular
        required: true
        type: string
      - description: Used when time type = 3 is customized
        in: query
        name: beginTime
        type: string
      - description: Used when time type = 3 is customized
        in: query
        name: endTime
        type: string
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/commun.MonitorChart2'
      summary: Query agent status
      tags:
      - Monitor
...
  /api/v1/user/login:
    post:
      parameters:
      - description: ' username '
        in: formData
        name: username
        required: true
        type: string
      - description: ' password '
        in: formData
        name: password
        required: true
        type: string
      summary: login
      tags:
      - user
...
  /api/v1/warn/getPushConfig:
    get:
      parameters:
      - description: ' smtp|wechat|dingtalk '
        in: formData
        name: name
        required: true
        type: string
      summary: getPushConfig
      tags:
      - Warn
...
  /api/v1/warn/savePushConfig:
    post:
      parameters:
      - description: ' object '
        in: body
        name: PushConfig
        required: true
        schema:
          $ref: '#/definitions/notice.PushConfig'
      summary: savePushConfig
      tags:
      - warn
  /api/v1/warn/saveThreshold:
    post:
      parameters:
      - description: ' object '
        in: body
        name: Threshold
        required: true
        schema:
          $ref: '#/definitions/handle.Threshold'
      summary: saveThreshold
      tags:
      - warn
  /api/v1/warn/warnList:
    get:
      parameters:
      - description: ' type '
        in: formData
        name: type
        required: true
        type: string
      - description: ' beginTime '
        in: formData
        name: beginTime
        required: true
        type: string
      - description: ' endTime '
        in: formData
        name: endTime
        required: true
        type: string
      summary: warnList
      tags:
      - warn
  /api/v1/warn/warnStatistics:
    get:
      parameters:
      - description: ' beginTime '
        in: formData
        name: beginTime
        required: true
        type: string
      - description: ' endTime '
        in: formData
        name: endTime
        required: true
        type: string
      summary: warnStatistics
      tags:
      - warn
swagger: "2.0"